Refactoring main screen.
