+++ source/android/app/src/main/java/com/github/jonforshort/androidlocalvpn/ui/main/TestTab.kt
@@ -24,90 +24,30 @@
 //
 // For more information, please refer to <https://unlicense.org>
 //
-package com.github.jonforshort.androidlocalvpn
+package com.github.jonforshort.androidlocalvpn.ui.main
 
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Send
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.runtime.*
-import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.ui.Alignment
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.Button
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.MutableLiveData
-import com.github.jonforshort.androidlocalvpn.ui.theme.AndroidLocalVpnTheme
-import com.github.jonforshort.androidlocalvpn.vpn.LocalVpnActivity
-import com.github.jonforshort.androidlocalvpn.vpn.isVpnRunning
-import com.google.accompanist.pager.ExperimentalPagerApi
-import com.google.accompanist.pager.HorizontalPager
-import com.google.accompanist.pager.pagerTabIndicatorOffset
-import com.google.accompanist.pager.rememberPagerState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import org.jsoup.Connection
 import org.jsoup.Jsoup
 import org.xbill.DNS.*
-import timber.log.Timber.d
-import timber.log.Timber.e
+import timber.log.Timber
 import java.io.IOException
 import java.net.SocketTimeoutException
 
-class MainActivity : LocalVpnActivity() {
-
-    private val vpnState = MutableLiveData(false)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            AndroidLocalVpnTheme {
-                Surface(color = MaterialTheme.colors.background) {
-                    MainView(
-                        tabData = listOf(
-                            "TEST" to Icons.Filled.Send,
-                            "SETTINGS" to Icons.Filled.Settings,
-                        ),
-                        onTabDisplayed = { index ->
-                            when (index) {
-                                0 -> {
-                                    TestTab()
-                                }
-                                1 -> {
-                                    SettingsTab(
-                                        vpnState.observeAsState(),
-                                        ::onVpnStateChanged
-                                    )
-                                }
-                            }
-                        }
-                    )
-                }
-            }
-        }
-
-        vpnState.postValue(isVpnRunning(this))
-    }
-
-    private fun onVpnStateChanged(vpnEnabled: Boolean) =
-        if (vpnEnabled) {
-            startVpn()
-        } else {
-            stopVpn()
-        }
-
-    override fun onVpnStarted() = vpnState.postValue(true)
-
-    override fun onVpnStopped() = vpnState.postValue(false)
-}
-
 @Composable
-private fun TestTab() {
+internal fun TestTab() {
 
     Column {
         TestHtmlQuery(
@@ -170,7 +110,7 @@ private fun TestHtmlQuery(text: String, url: String) {
                 val html = resp.body()
                 val duration = System.currentTimeMillis() - requestStartTime
 
-                d(
+                Timber.d(
                     """
                         |dumping html, count=[${html.length}] duration=[$duration]
                         |$html
@@ -179,13 +119,13 @@ private fun TestHtmlQuery(text: String, url: String) {
                 )
                 onRequestFinished(true)
             } catch (e: SocketTimeoutException) {
-                e(e, "Request timed out")
+                Timber.e(e, "Request timed out")
                 onRequestFinished(false)
             } catch (e: IOException) {
-                e(e)
+                Timber.e(e)
                 onRequestFinished(false)
             } catch (e: RuntimeException) {
-                e(e)
+                Timber.e(e)
                 onRequestFinished(false)
             }
         }
@@ -231,7 +171,7 @@ private fun TestDnsQuery(text: String, domain: String, server: String = "8.8.8.8
                     .whenComplete { answer, e ->
                         if (e == null) {
                             val duration = System.currentTimeMillis() - requestStartTime
-                            d(
+                            Timber.d(
                                 """
                                 |dumping dns, duration=[$duration]
                                 |$answer
@@ -240,14 +180,14 @@ private fun TestDnsQuery(text: String, domain: String, server: String = "8.8.8.8
                             )
                             onRequestFinished(true)
                         } else {
-                            e(e)
+                            Timber.e(e)
                             onRequestFinished(false)
                         }
                     }
                     .toCompletableFuture()
                     .get()
             } catch (e: Exception) {
-                e(e)
+                Timber.e(e)
                 onRequestFinished(false)
             }
         }
@@ -273,82 +213,5 @@ private fun TestDnsQuery(text: String, domain: String, server: String = "8.8.8.8
         }
     ) {
         Text(text)
-    }
-}
-
-@Composable
-private fun SettingsTab(
-    isVpnEnabled: State<Boolean?> = mutableStateOf(false),
-    onVpnEnabledChanged: (Boolean) -> Unit = {}
-) {
-    Row(
-        modifier = Modifier
-            .height(IntrinsicSize.Max)
-            .padding(10.dp)
-            .fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween
-    )
-    {
-        Text(
-            text = "Enable VPN",
-        )
-        Switch(
-            checked = isVpnEnabled.value ?: false,
-            onCheckedChange = { onVpnEnabledChanged(it) },
-        )
-    }
-}
-
-@OptIn(ExperimentalPagerApi::class)
-@Composable
-private fun MainView(
-    tabData: List<Pair<String, ImageVector>> = emptyList(),
-    onTabDisplayed: @Composable (index: Int) -> Unit = {}
-) {
-    val pagerState = rememberPagerState()
-    val tabIndex = pagerState.currentPage
-    val coroutineScope = rememberCoroutineScope()
-    Column {
-        TabRow(
-            selectedTabIndex = tabIndex,
-            indicator = { tabPositions ->
-                TabRowDefaults.Indicator(
-                    Modifier.pagerTabIndicatorOffset(pagerState, tabPositions)
-                )
-            }
-        ) {
-            tabData.forEachIndexed { index, pair ->
-                Tab(selected = tabIndex == index, onClick = {
-                    coroutineScope.launch {
-                        pagerState.animateScrollToPage(index)
-                    }
-                }, text = {
-                    Text(text = pair.first)
-                }, icon = {
-                    Icon(imageVector = pair.second, contentDescription = null)
-                })
-            }
-        }
-        HorizontalPager(
-            state = pagerState,
-            modifier = Modifier.weight(1f),
-            count = tabData.size
-        ) { index ->
-            Column(
-                modifier = Modifier.fillMaxSize(),
-                verticalArrangement = Arrangement.Top,
-                horizontalAlignment = Alignment.Start
-            ) {
-                onTabDisplayed(index)
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DefaultPreview() {
-    AndroidLocalVpnTheme {
-        MainView()
     }
 }
